pyproject.toml
[project]
name = "gg-bench"
version = "0.1.0"
description = "Graph compression benchmark for evaluating LLM effectiveness at graph compression tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.1.1",
    "networkx>=3.0",
    "rich>=14.1.0",
    "flask>=3.1.2",
    "pandas>=2.3.2",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "numpy>=2.3.2",
    "tqdm>=4.66.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.10",
    "ty>=0.0.1a19",
]

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]
